+++ vectortiles/backends/postgis/__init__.py
@@ -54,4 +54,5 @@ class VectorLayer(BaseVectorLayerMixin):
                 ],
             )
             row = cursor.fetchone()[0]
-            return row.tobytes() if row else b""
+            # psycopg2 returns memoryview, psycopg returns bytes
+            return row.tobytes() if type(row) == memoryview else row or b""