+++ docs/changelog.rst
@@ -24,6 +24,10 @@ CHANGELOG
 
   * Add DRF and MapLibre examples
 
+* Legacy
+
+  * Unpin protobuf version as latest mapbox-vector-tile version is fixed.
+
 
 0.2.0       (2022-10-17)
 ------------------------
@@ -64,7 +68,6 @@ First beta release
 ------------------------
 
 # First Release
-
   * Generate Vector Tiles from django models
       * in python
       * with PostGIS